changed deep copy to jackson instead of toBuilders
